requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
#poloniex==0.4.6
psutil==5.4.1
pymongo==3.5.1
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
